requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
groq==0.25.0
hindsight-client>=0.1.0
python-dotenv==1.0.1
